backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.3
ultralytics==8.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
tensorflow==2.15.0
onnxruntime==1.16.3
pydantic==2.5.0
aiofiles==23.2.1

